Docker and Kubernetes Deployments
Adding Dradis to a container is useful for a standalone process.
Building in Jenkins or spinning up for note taking, the Dockerfile adds value.
The kubernetes deployment is useful for a hosted service where the client does
not have resources to run Dradis. Hosting on a pod consumes less resources and
containerizes the Dradis Framework.